backend/requirements.txt
# Core ML, Data Handling, and Utilities
pandas
numpy==2.1.3
scikit-learn==1.6.1
joblib
tensorflow==2.19.0
python-dateutil
lightgbm
matplotlib
seaborn

# Deep Learning for Text Model (Hugging Face)
transformers==4.41.2
safetensors==0.4.3
torch==2.3.1
datasets

# Gradient Boosting for Network Model
xgboost

# PE File Analysis
pefile

# API Framework
fastapi
uvicorn
sqlalchemy

# NLP (for the Sensitive Data Classifier model)
spacy

# For making HTTP requests (downloading data, testing API)
requests

python-multipart

# For generating test data
Faker

# For JWT authentication (future use)
python-jose[cryptography]
# firebase
firebase-admin
google-cloud-storage

passlib[bcrypt] 
